+++ packages/cubejs-cubestore-driver/codegen/generate.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-flatc --ts ../../../rust/cubestore/cubestore/src/codegen/http_message.fbs --ts-flat-files
+flatc --ts ../../../rust/cubeshared/src/codegen/http_message.fbs --ts-flat-files
 mv http_message.ts index.ts